Refactored user login/signup/logout/management into account blueprint
