+++ app/templates/shared/template.html
@@ -21,6 +21,7 @@
     <div id="left_gutter"></div>
 
     <div id="content_container">
+        Current User: {{ current_user }}
         {% block content %}
         {% endblock %}
     </div>
